quantize_apply code with new API wrapper
Pulls the quantize_annotate/quantize_apply functions out
into new files. Modifies the code to use the new wrappers,
use registry and QuantizeProviders.

Also modifies and adds tests accordingly.

PiperOrigin-RevId: 265757533
