+++ tensorflow_model_optimization/python/core/quantization/keras/quantize_emulate.py
@@ -85,105 +85,3 @@ def QuantizeEmulate(to_quantize,
     return keras.models.Sequential(_QuantizeList(to_quantize.layers, **params))
   elif isinstance(to_quantize, keras.layers.Layer):
     return QuantizeEmulateWrapper(to_quantize, **params)
-
-
-# TODO(pulkitb): Enable lint naming is fixed and made consistent.
-def quantize_annotate(to_quantize, **kwargs):  # pylint: disable=invalid-name
-  """Specify a layer or model to be quantized.
-
-  This function does not apply an quantization emulation operations. It merely
-  wraps the keras layer (or each layer in the model) with `QuantizeAnnotate`
-  to note which layers need to be quantized.
-
-  Args:
-    to_quantize: Keras layer or model to be quantized.
-    **kwargs: Additional keyword arguments to be passed to the keras layer.
-
-  Returns:
-    Keras layer wrapped with `QuantizeAnnotate` if layer is passed. Else,
-    a new keras model with each layer in the model wrapped with
-    `QuantizeAnnotate`.
-  """
-
-  def _add_quant_wrapper(layer):
-    if isinstance(layer, quant_annotate.QuantizeAnnotate):
-      return layer
-
-    return quant_annotate.QuantizeAnnotate(layer)
-
-  if isinstance(to_quantize, keras.Model):
-    return keras.models.clone_model(
-        to_quantize, input_tensors=None, clone_function=_add_quant_wrapper)
-  elif isinstance(to_quantize, keras.layers.Layer):
-    # TODO(pulkitb): Since annotation for model and layer have different
-    # parameters, we should likely remove support for layers here.
-    return quant_annotate.QuantizeAnnotate(to_quantize, **kwargs)
-
-
-def quantize_apply(model):
-  """Apply quantization operations to a keras model.
-
-  This function takes a keras model which has been annotated with
-  `quantize_annotate` and constructs a new keras model in which each of the
-  annotated layers have been quantized. The quantization process introduces
-  new quantization ops in the Tensorflow graph to appropriately emulate
-  quantization loss.
-
-  Note that to exactly emulate quantization loss, certain graph/model
-  transformations may be applied. This is required since the actual quantized
-  kernel implementations may apply similar transformations.
-
-  Args:
-    model: A keras Sequential or Functional model which has been annotated
-    with `quantize_annotate`.
-
-  Returns:
-    Returns a new cloned keras model in which the annotated layers have been
-    quantized. All the existing layers are cloned.
-  """
-
-  if not isinstance(model, keras.Model):
-    raise ValueError('Only a keras `Model` instance can be used.')
-
-  if not isinstance(model, keras.Sequential) \
-      and not model._is_graph_network:  # pylint: disable=protected-access
-    raise ValueError('model should be either a keras.Sequential or a '
-                     'keras functional model.')
-
-  # Have at least 1 layer annotated with QuantizeAnnotate
-  if not any(isinstance(layer, quant_annotate.QuantizeAnnotate)
-             for layer in model.layers):
-    raise ValueError('model does not contain any layers which have been '
-                     'annotated with `quantize_annotate`. There are no layers '
-                     'to quantize.')
-
-  if not model.built:
-    raise ValueError('quantization cannot be applied to a model which has not'
-                     'been built yet. Please call `model.build(input_shape)`'
-                     'before quantizing your model.')
-
-  def _clone_model_with_weights(model_to_clone):
-    cloned_model = keras.models.clone_model(model_to_clone)
-    cloned_model.set_weights(model_to_clone.get_weights())
-
-    return cloned_model
-
-  def _apply_quantization(quant_annotate_layer):
-    return QuantizeEmulateWrapper(
-        quant_annotate_layer.layer,
-        **(quant_annotate_layer.get_quantize_params()))
-
-  # Create a copy of the model with the same weights. We can then quantize this
-  # model without modifying the weights of the original model.
-  model_copy = _clone_model_with_weights(model)
-
-  def _add_quant_emulate_wrapper(layer):  # pylint: disable=missing-docstring
-    if not isinstance(layer, quant_annotate.QuantizeAnnotate):
-      return layer
-
-    # Use QuantizeEmulate wrapper on annotated layer which actually
-    # quantization ops.
-    return _apply_quantization(layer)
-
-  return keras.models.clone_model(
-      model_copy, input_tensors=None, clone_function=_add_quant_emulate_wrapper)
